Refactor Bootstrap.Wapper.build_tag/1
With Phoenix.HTML.raw/1 we can render a empty tag and it helps when we
need render or not a tag given a condition. I'm rafactoring the code to
use this in errors_tag/1 and hint_tag/1 because they are going to be
rendered only when their attribute are present.